I meant to make another commit yesterday, but I got sidetracked. Now that I am calculating the costs properly and have set the learning rates I am set to start analyzing why the mi_hebb_prong layer does not work as it should. Here is how it runs.
```
The learning rate is 2 ** -13.50531
Starting timing for: iteration 0
The time was 00:00:16.2598687 for: iteration 0
Training: 4.45427812370882
The learning rate is 2 ** -13.50531
Starting timing for: iteration 1
The time was 00:00:16.7788989 for: iteration 1
Training: 3.04306093376876
The learning rate is 2 ** -13.50531
Starting timing for: iteration 2
The time was 00:00:16.0529950 for: iteration 2
Training: 2.65367713252336
The learning rate is 2 ** -13.50531
Starting timing for: iteration 3
The time was 00:00:16.2911295 for: iteration 3
Training: 2.48646558441396
The learning rate is 2 ** -13.50531
Starting timing for: iteration 4
CuSolve function failed!
The leading minor of order 85 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:679: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
```

It is really, really strange that this blows up once it gets past 2.4. It does that regardless of the learning rate. What makes it eerie is that it trains well up to then, just as good as a regular layer.

This really reminds me of that sampling function failure that I fixed a few days ago. Of course, there is not chance that this is the same bug.

But it is the same kind of bug no doubt.

There is something I am suspecting without any rational basis for doing so.

`n * x + (one - n) * H`

I am suspecting the `n` here. The way this equation is set up makes me believe that n should be in the (0,1) range, but in fact there is absolutely nothing preventing from slipping out of there.

The first thing I have to try is just making n a hyperparameter.

For the kind of thing that `n` is I might as well go that route. I do not think I'll be getting much benefit in keeping it tied.

After thinking about it for a while it is clear that the way Hebbian plasticity would be great for doing credit assignment over long time horizons. Since the sequence length I am testing it on is only 64, I might as well set n to 1/64.

Let me give it a shot. Right now I am wondering whether `n` in Oja's rule should also be in the (0,1) range. I'll leave that mystery for later.

Also when I am done with this I am going to dedicate some time to properly replying to theindigamer.