"9:15am.
```
The learning rate is 2 ** -12.49694
Starting timing for: iteration 0
The time was 00:00:13.6437863 for: iteration 0
Training: 3.78916834829793
The learning rate is 2 ** -12.49694
Starting timing for: iteration 1
The time was 00:00:13.5002925 for: iteration 1
Training: 2.59734936611194
The learning rate is 2 ** -12.49694
Starting timing for: iteration 2
The time was 00:00:13.7062501 for: iteration 2
Training: 2.36060137229823
The learning rate is 2 ** -12.49694
Starting timing for: iteration 3
The time was 00:00:14.5025087 for: iteration 3
Training: 2.20373454972479
The learning rate is 2 ** -12.49694
Starting timing for: iteration 4
The time was 00:00:14.9670802 for: iteration 4
Training: 2.08427414213036
```

Now it works. My intuition was correct in fact. I've also cut the running time by 1/4th.

Let me lower the learning rate a bit more.

```
The learning rate is 2 ** -11.50112
Starting timing for: iteration 0
The time was 00:00:16.0413879 for: iteration 0
Training: 3.22564496833023
The learning rate is 2 ** -11.50112
Starting timing for: iteration 1
The time was 00:00:15.3887020 for: iteration 1
Training: 2.3062847312114
The learning rate is 2 ** -11.50112
Starting timing for: iteration 2
The time was 00:00:15.5895556 for: iteration 2
Training: 2.07116087080682
The learning rate is 2 ** -11.50112
Starting timing for: iteration 3
The time was 00:00:15.2669854 for: iteration 3
Training: 1.9105494610653
The learning rate is 2 ** -11.50112
Starting timing for: iteration 4
The time was 00:00:16.0063291 for: iteration 4
Training: 1.78162513876785
```

Complete stability is what I am getting.

```
The learning rate is 2 ** -10.49903
Starting timing for: iteration 0
The time was 00:00:14.9007184 for: iteration 0
Training: 2.79806285918826
The learning rate is 2 ** -10.49903
Starting timing for: iteration 1
The time was 00:00:14.5286565 for: iteration 1
Training: 2.05362753128683
The learning rate is 2 ** -10.49903
Starting timing for: iteration 2
The time was 00:00:14.6396367 for: iteration 2
Training: 1.83692435138022
The learning rate is 2 ** -10.49903
Starting timing for: iteration 3
The time was 00:00:14.5767775 for: iteration 3
Training: 1.70023951553093
The learning rate is 2 ** -10.49903
Starting timing for: iteration 4
The time was 00:00:14.6243695 for: iteration 4
Training: 1.61051399645939
```

Higher, I need to go higher. After that I'll compare it to standard MI.

But I think these results are better than what I was getting with MI already.

```
The learning rate is 2 ** -9.500073
Starting timing for: iteration 0
The time was 00:00:16.2608461 for: iteration 0
Training: 2.49917191926076
The learning rate is 2 ** -9.500073
Starting timing for: iteration 1
The time was 00:00:16.3407826 for: iteration 1
Training: 1.96221696008765
The learning rate is 2 ** -9.500073
Starting timing for: iteration 2
The time was 00:00:16.3240077 for: iteration 2
Training: 1.87534513286957
The learning rate is 2 ** -9.500073
Starting timing for: iteration 3
The time was 00:00:16.3464076 for: iteration 3
Training: 1.88686776532179
The learning rate is 2 ** -9.500073
Starting timing for: iteration 4
The time was 00:00:15.8612030 for: iteration 4
Training: 1.87963295750889
```

The training is starting to asymptote. Let me try increasing it to -9 and then -8.5 and then I will move to the `mi_prong`.

```
The learning rate is 2 ** -9.000094
Starting timing for: iteration 0
The time was 00:00:16.1843044 for: iteration 0
Training: 2.43463861883058
The learning rate is 2 ** -9.000094
Starting timing for: iteration 1
The time was 00:00:15.9813532 for: iteration 1
Training: 2.13984197705547
The learning rate is 2 ** -9.000094
Starting timing for: iteration 2
CuSolve function failed!
The leading minor of order 4 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:674: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
```

-9 is just too high. This is good enough.

```
The learning rate is 2 ** -9.999356
Starting timing for: iteration 0
The time was 00:00:11.1194848 for: iteration 0
Training: 2.63817095877697
The learning rate is 2 ** -9.999356
Starting timing for: iteration 1
The time was 00:00:10.9980651 for: iteration 1
Training: 1.96779351476682
The learning rate is 2 ** -9.999356
Starting timing for: iteration 2
The time was 00:00:11.0194847 for: iteration 2
Training: 1.77882689541819
The learning rate is 2 ** -9.999356
Starting timing for: iteration 3
The time was 00:00:11.0270062 for: iteration 3
Training: 1.69068087242506
The learning rate is 2 ** -9.999356
Starting timing for: iteration 4
The time was 00:00:11.2093121 for: iteration 4
Training: 1.65442746834303
```

Here is the MI+PRONG layer.

```
The learning rate is 2 ** -9.500073
Starting timing for: iteration 0
The time was 00:00:11.8491181 for: iteration 0
Training: 2.50635309358072
The learning rate is 2 ** -9.500073
Starting timing for: iteration 1
The time was 00:00:12.0107859 for: iteration 1
Training: 1.95464366367426
The learning rate is 2 ** -9.500073
Starting timing for: iteration 2
The time was 00:00:11.8662624 for: iteration 2
Training: 1.85806370714862
The learning rate is 2 ** -9.500073
Starting timing for: iteration 3
The time was 00:00:12.0716096 for: iteration 3
Training: 1.86284106496953
The learning rate is 2 ** -9.500073
Starting timing for: iteration 4
The time was 00:00:12.1304489 for: iteration 4
Training: 1.87031843250314
```

They are equal on the char RNN. I am not seeing any benefit from using Hebbian plasticity, but maybe that is just the wrong thing for the job here. The MI RNN works well already. It is important that a metalearning mechanism does not get in the way when it is not useful.

All in all, this little session was a dreamlike one. I guessed the exact spot where the thing is breaking and fixed it on the first try.

Let me factor out the `n` as a hyperparameter...done.

9:50am. I am thinking what next is on the agenda. I am really curious about Oja's rule. I want to understand what it is doing as it is not at all like Hebbian plasticity.

I'll email Miconi to ask him about that amongst other things. My first email was a miss - it seems I was not clear enough.

Let me deal with the Github issue first."