Time to get back into the action.
I am thinking what comes next right now. I'll work on that email, but just to be sure I think I should compare a MIRNN on one of the metalearning tasks.

If the plastic MIRNN worked much better than the standard one then that would be it, but I am not convinced just yet that I did it all correctly. I need to do more testing in order to make sure.

Let me get that email out of the way. I am not sure whether there is a point to it, but I should just follow the steps I've outlined for myself.