The conversation I am having with over email with Miconi reminds me of the one I am having with theindigamer, except the positions are reversed.
Right now, I am the one getting confused about differentiable plasticity. I was thinking of it in terms of it being an RNN, but it seems that is wrong.

http://www.scholarpedia.org/article/Hopfield_network
http://www.scholarpedia.org/article/Oja_learning_rule

I think that just like I did with natural gradients, I am going to stop my programming frenzy and educate myself for a while.

Thanks to that second link now I understand what Oja's rule is doing. I did not think it was connected to the first equation, but it is.

Somehow a picture is coming in my mind, but so do questions. If Hebbian plasticity is doing PCA, then how does that relate to PRONG? PRONG also does something similar to Hebbian learning except it uses it for optimizing the net.

It occurs to me now that I haven't completed the mi_hebb architecture. The inputs should also have plastic weights.

Since Oja is the one who wrote the PCA book, now I finally have a reason to take a bite into it. Also I am going to take a look at that Brain course I saw a while ago.

...I think I am starting to understand what picture exactly is trying to enter my mind.

Think about what the global learning rates for the entire network would be and what the local learning rates for plastic connections would be.

The local learning rates would be orders of magnitude faster.

Even so...how would I think of a feedforward net with plastic connections if it not in terms of it being a recurrent network?

Maybe I could think of it creating a zone of flexibility where the network is capable of doing localized short-term learning?