I got up 45m ago and have been chillnig. Enough Reddit posting and asking and answering questions. It does not really matter. I've been overdoing it for the past few days in that regard.
What I need to do is keep pushing forward. I said I would stop the programming frenzy to do studying yesterday, but that is the wrong choice. I need to keep the pace going.

I will have plenty of time to study on the side when the agents are training. I absolutely must break through to that segment of my journey. I'll do the studies for the next step after that.

Extra information will not help me at this point. I need to get the pieces together.

Today my goal will be to do the binary memorization test.

That will tell me quite clearly whether plasticity is doing anything or not - that is whether my implementation is correct. Once I get my confirmation I return to doing reinforcement learning.

http://www.cs.cmu.edu/~bhiksha/courses/deeplearning/Fall.2016/pdfs/OJA.pca.pdf
https://www.reddit.com/r/MachineLearning/comments/9d81oz/d_has_there_been_any_work_on_an_apibased_agi/

Let me read this paper and that thread and then I will start work. The test won't be at all hard to do.