Ahhh...damn this all to hell.
I just trying to fit this all in my head and fells like it is slipping. There is just so much complexity here that I feel overwhelmed.

By the time it gets to the final branch I will have forgotten what is what.

Agghhhh...

This can't be the best I can do.

I need to focus more. I knew the loop would be troublesome. I'll figure out a way to deal with it eventually.