The training loop broke me it seems...
I do not think I will doing any more programming for the day so let me collect my thoughts.

There are total of 5 training loops scattered throughout the library. 2 of the are for the PG and the AC poker player which use `mutable_function` and the other 3 are each in the `learning` tests.

I think my issue here is this - I understand imperative programming and I understand functional programming.

Now that I am reevaluating my programming life, I am not that sure that I understand dynamic imperative programming. Deep down I am perplexed.

I mean the day started out well enough as planned. It did take me an hour to figure out how I want to make the dataset. Making the masks and putting in the shuffle functions and loading it on the GPU did take a bit of time.

But Jesus these loops, these damn loops! When I was doing `learning2` the training loop also took me an insane amount of time, like the whole day to just make it. And `learning1` was hard as well.

And the two loops for the poker players were also insanely hard.

While I was doing them and now this one the one thought that keeps recurring in my mind over and over is - 'I wish Spiral was a dynamic language so I did not have to deal with this bullshit. I do not care about performance here anyway.'

This shouldn't possibly be this hard. None of this.

I am certain that making the loop for the Binary Memorization test should not have taken more than 15 minutes.

...

This is not the first time I've been in situation such as this and it is only my arrogance that was starting to think there is nothing left to learn. But it also arrogance to think that there is anything more to learn. Humility is a trap - if there is something more to learn then there is pressure to reply to the question of what that is. Nobody can live up to that standard and that soul crushing fakeness.

There is only the present moment and the zone. The abyss lies beyond that.

Forget machine learning for a moment. A bigger battle is at play here.

Every time I've been in a similar situation I've surmounted it by letting go of my preconceptions, forgetting what is and sketching out what it is that I want. Then I went towards it. So I will do that here.

After I am done with this, these loops will never bother me again. I am the one who controls the code, not the other way around.

Real life sucks so much - all the games the normies play are so low stakes and pointless. They compete over nothing. But programming as a game has promise. Now that the celestial and terrestrial are linked by a thread all I need to do is go forward and complete the bridge.

This is not Spiral's true power. I went trully far with pure functional programming and CPS where it was warranted, but one technique cannot do everything even if it is quite powerful. I admit to thinking that there is no point to dynamic typing in the past. It seems that going from here I am going to have another look at that.

Spiral's union types are not like ML's. I can do so much more with them. I can surmount this.