``` met train {data network learning_rate final} s = // TODO: Work in progress.     inl cost = ref 0.0
    loop_over size.seq <| inl i ->
        inl data = index_into i data

        loop_over size.shot <| inl i ->
            loop_over size.pattern_repetition <| inl _ ->
                network.run (data.original (order i))

            loop_over size.empty_input_after_repetition <| inl _ ->
                network.run zero

        inl i = random size.pattern
        network.run (data.degraded i)
        network.peek <| inl inl {final} -> cost := cost() + final (data.original i)
        network.pop_bcks {learning_rate=learning_rate ** 0.85}
        network.optimize learning_rate
```

When I said I wanted a loop, this is what I had in mind. Just what is the thing that I wrote previously? It is nothing. 100 lines of garbage.

Damn, I am tired. I'll stop here for the day and move to implementation tomorrow.

Once that is done, I'll replace the 5 loops that exist with a single mechanism. That will be the final proof of my understanding. Never again will this cause me to stumble.