"8:20am. I did some thinking about dynamic types. If I had to implement them I know how I would do it. But they are complicated beasts when it comes to optimizing them.
Yeah, it is complicated all together. Adding dynamic types would add a significant amount of complexity to the language.

It is something to be avoided. I wonder if I will ever encounter a situation where using them would be unavoidable? Probably not.

8:40am. I do not feel like starting right now. Let me play around for a while.

I do not understand it myself. Though the techniques I will be using to make the sketch should be obvious why are they such a complete mental blind spot to me? Am I really that averse to anything even remotely involving dynamism? Or is it because they require a different kind of reasoning to which I am not used to?

But without a doubt the `mutable_function` is exactly the right solution for using NN agents in the poker game. It might be in theory be possible to thread the state through the game and CPS everything, but in practice that is absolutely out of the question. I cannot change an entire game every time I make a change to the NN agent! Threading state introduces coupling.

In small doses coupling is good as increases the coverage that the type system provides, but too much of it just gets in the way.

As an example - in that loop yesterday it is much harder to thread state through multiple loops that it is to just to use a reference. When it is 1-2 loops it is not a problem, but with 5 nested loops I was already starting to lose track of where the brackets should be indented, let alone what the state is supposed to be.

I need to dwell on this for a bit. The ultimate goal is to make better code, rules are just a means not an end for that.

Branches with dynamic returns, state-bound functions - I think as I go forward from here I should keep my mind open to them.

I need to abstract away the NN agents completely and the same goes for these loops."