Done with breakfast. Let me resume.
```
        map=inl {network s} input ->
            inl network, input = run s input network
            inl final label =
                inl {out bck} = final label input s
                bck()
                s.CudaTensor.get out |> to float64
            (), {network s final}
```

You know after all I'll go back to the usual thing. Since it does boxing of outputs it would be just too much trouble to do it with tuples...

...You know, one library change that might be welcome could be to allow tuples to operate on singletons. There is too much back and forth with Tuple.wrap and Tuple.unwrap.

Forget this for now. That was something I could have done while making the Tuple module long ago. I am done with those redesigns. Now the priority is to make full use of Spiral.

```
        map=inl {state={network s} input} ->
            inl network, input = run s input network
            inl final label =
                inl {out bck} = final label input s
                bck()
                s.CudaTensor.get out |> to float64
            {state={network s final}}
```

Let me do it like this. It was how it was originally.

Hmmmm...well, I think I have enough to start thinking about moving to testing. I'll probably move the `runner` into the ML library after that.