"1:15pm. Thunder. I'll shut down.
3:15pm. I was in bed all this time. I think it slowed down now. It was just as I was getting into things. Damn weather.

But I think some deliberation did me good. I am still reeling a bit from yesterday. I think I will find a new way of thinking at the end of all of this.

```
Error trace on line: 44, column: 13 in file "Loops".
            if check from then
            ^
Types in branches of If do not match.
Got: {} and []
```

I know why this is happening.

```
        | k input ->
            inl {d with state} = (funs k) {state=states.last; input}
            states.add state
            {d without state}
```

Because of the return here.

I'll be frank maybe I should in fact redesign this a little. Why shouldn't the return of the function be everything?

```
        | k input ->
            (funs k) {state=states.last; input}
            |> states.add
```

Now that I have more flexibility in bck, I should take advantage of it.

```
        map=inl {state={network s} input} ->
            inl network, input = run s input network
            inl final label =
                inl {out bck} = final label input s
                bck()
                s.CudaTensor.get out |> to float64
            {network s final}
```

Let me do it like this. And I will change the way inputs are done after all.

In the end, new techniques lead to new ways of programming.

I've been stupidly glued to threading state everywhere like a functional monkey.

I knew that Spiral had good OO capabilities and I've been taking advantage of it, but there is a level beyond the one I've been used to."