I think I will call it a day here.
The thing just is not working. I am quite confused right now and have no idea where the problem could be.

I think that it is here that the real work on differentiable plasticity will start.

I am going to have to dive into the Python code and study it carefully, like by line. Unless something jumps out at me tomorrow, I will need to install PyTorch to get it to run.

It is just...it is confusing that differentiable plasticity is doing nothing for the MI layer.

But beyond that I am a bit confused how to do the test based on how it was described in the paper.

Are the episodes supposed to be repated or not?

But even ignoring that, I can immediatelly understand that this task is actually a very hard one, at least compared to the char RNN.

I underestimated it at first. It does involve some very long term credit assignment.

If I can get the RNNs to master this one, I will be a lot better prepared going into the future. I can only thank Miconi and friends for preparing this for me.

This will take a while it seems. Rudder is still ways off.

Well, forget that thing for now.

Plasticity is definitely a vital ingredient for my future success. Even if I take a month playing with it that would be worth the price.

This bring me back to 2015 when I was making that sparse autoencoder in F#. I was just picking up ML for the first time back then. I've come a long way since then.

I'll keep going at it. Without a doubt, I will be finished with this phase by the end of the year.