I am studying the code on the Uber Github and am not seeing that much difference between theirs and my own.
The way they create the inputs is just so attrocious though. I'd honestly flunk whoever was writing the input and target generation.

I think I get something now...the episodes are not repeated and go on for 2k iterations.

There are some notable exceptions where they strenghten the inputs, and give the test example for multiple steps, but that would be it. Also they do use the biases, but say that does not matter.

There are differences in initalization, but nothing that I would guess would cause the test to fail.

...Once i run it I will know. I'll simplify how the generation of inputs and labels is done along the way. I'll step through it every step until I see how it is done.

I guess this is a good opportunity to warm up to Python a little. I've bashed it every step of the way, but I should know how to take advantage of it better now.

Let me do an extra bit of work so I do not have to do it tomorrow as well. I'll print out the inputs the network gets just so I can make sure that everything is order.