Let me chill a bit until my brain gets online and then I will start. Today I will be installing PyTorch, cloning the differentiable plasticity repo and getting the Binary Pattern test to run.
After that I'll go between the two tests until they are equal. Once I get Spiral's test is on an even footing with the other one and gets the same results that will allow me to understand where I went wrong. Right now there are just too many differences to be sure of anything.

I suppose it is always possible that I have an implementation error in the `hebb` layers, but it is more likely the case that at least one of my assumptions is broken.