Enough slacking. Let me just do it. I need to do the necessary research. The project won't complete itself on its own.
Yesterday I tried to install PyTorch but got the same kind of error as when I installed Tensorflow back when I was working on relative natural gradient. I forgot how I managed to get around it that time, but I think I found the answer on SO. Well, I'll figure it out somehow. It is time to start this. For now the goal will be to just run that thing successfully.