This might be the smoothest installation in my life. I just installed it, removed the dead enviroments and it all just works the first time as it should. That is a miracle.
Years ago these DL libraries were such a massive pain in the ass and now they are easy to use.

If I am really serious about making Spiral popular maybe I should just merge it with Python. Make a version of it that compiles to C++ and make it callable as a library in Python. That would get some eyeballs.

What I am doing right now on the PL sub and elsewhere is a complete waste of time, though I admit that given the state of Spiral right now that I'd at least have gotten one user apart from myself. Especially after that HN link.

I am a fuckup in a way. It goes all the way back to my trading days.

If I had just compromised back then and gotten a job in order to make stake so that I could trade comfortably the endeavor might have very well been a success. Especially considering that I know several ways of trading that would have worked.

The reason why I did not do that is because trading was a way for me to avoid having to go down that way in the first place.

Now with Spiral I have a choice of trying to draw in users. If I really want that I should merge it with Python - a dynamic language like it would do a lot better job of being the glue than F#. After that I should try and impress the TF and the PyTorch guys. Maybe do a better job of the documentation so it does not feel like a salesman's pitch. Give it an aura of normality. If I went down that route and did the sane thing for a change I could probably angle in a job at Google or Facebook working on languages.

That is not a small thing from a monetary perspective given what sort of salaries they offer. I could work on that for a few years and then go into trading and whatever after that.

My programming talents are not quite the best in the world - even I think I am too dumb for that to be the truth, but they are definitely significantly better than the average pro programmer. I could go far with it on the normal route.

Make a lot of money, invest in stocks, do well, do very well. Go through life smoothly.

That sort of outcome would be most undesirable. How could it possibly be acceptable?

Once the Singularity starts I'd honestly die of embarassment (and then shortly die for real) if I ever found myself in the situation of being a rich normie. The wave of normality is the false wave. I want to become one of the Inspired even if I have to sacrifice everything else.

It is just difficult to put into words, but what I wrote in the story Simulacrum is just the tip of the iceberg.

I've had seen it described as a story of singular grizzliness, but that is not it. There is no way my imagination is good enough in the first place to imagine singular grizzliness. Self improvement as described there is a day in an amusement park. An easy counter point to scheme proposed in the story is that it is based on evolution. Evolution is inefficient. Too inefficient to be a real part of anything of importance. So one needs to look elsewhere for a break.

The real question is how would a story of self improvement based on backprop look like. I have some ideas, but it is just so outlandish and unacceptable. I need to collect more pieces to make the picture make sense. I need to do it as I will be depending on it to trigger the process in the real world after all. Right now I barely understand what differentiable programming is.

Also regardless of the current status of the backprop perspective, it lacks the element of fear. The true self improvement process should be a black maw, an absolute terror and a void so deep that it chills the soul. Right now the way the picture looks is hope itself and I cannot accept this.

I guess this is how I changed along the way somewhere. When I was younger I justifiably cared about power, but now I do not care about it in that sense. I understand that all paths lead to the pinnacle.

Instead what is really interesting is the price of power.

...I do not know where the last hour went. Let me go back to playing with PyTorch. If it is not other people distracting me, I will find some way to do it myself. But this sort of thing is important for stabilizing my motivation so it is fine in the end. Even if I want to program 10h a day, no way will that happen.