"1:15pm.
```
    to = target.cpu().numpy(); yo = y.data.cpu().numpy()[0][:PATTERNSIZE]; z = (np.sign(yo) != np.sign(to)); lossnum = np.mean(z)  # Saved loss is the error rate

    total_loss  += lossnum
```

What is this?

```
(1999, '====')
[ 1. -1. -1.  1.  1. -1. -1. -1.  1.  1.]
[-20.  -0.   0.  20. -20.  -0. -20.  20.   0.  20.]
[-1.    -0.776 -0.335  1.    -1.    -0.842 -1.     1.    -0.064  1.   ]
Time spent on last 10 iters:  0.1874842643737793
Mean loss over last 10 iters: 8.825773811340332
```

Ok, this is how it does with the usual loss. This is with a pattern size of 50.

```
(229, '====')
[ 1. -1.  1. -1.  1. -1. -1.  1.  1.  1.]
[-20.  20.  -0.   0. -20.  -0.   0.   0.   0.  20.]
[-1.     1.    -0.881  0.953 -1.    -0.957 -0.11   0.826  0.839  1.   ]
Time spent on last 10 iters:  2.858896255493164
Mean loss over last 10 iters: 74.21193370819091
```

It does a lot better on larger patterns.

It is easy to understand why that might be. Small patterns might have a lot of overlap so the network might have a hard time training. Unlike my own network though, this one does rather well regardless of the size. My can't do better than random.

The next step would be to replace the Adam optimizer with an SGD one."