"1:55pm.
```
inputs:
[|
    [|-1; 1; -1; -1; -1; 1|] // 1
|]

[|
    [|-1; 1; -1; -1; -1; -1|] // 2
|]

[|
    [|1; -1; 1; -1; 1; -1|] // 3
|]

[|
    [|1; -1; -1; -1; -1; 1|] // 4
|]

[|
    [|1; -1; -1; -1; -1; 1|] // 4
|]

[|
    [|1; -1; -1; -1; -1; 1|] // 4
|]

[|
    [|-1; 1; -1; -1; -1; 1|] // 1
|]

[|
    [|1; -1; -1; -1; -1; 1|] // 4
|]

[|
    [|1; -1; 1; -1; 1; -1|] // 3
|]

[|
    [|-1; 1; -1; -1; -1; -1|] // 2
|]

[|
    [|0; -1; 1; -1; 0; 0|]
|]

target:
[|
    [|1; -1; 1; -1; 1; -1|] // 3
|]

Cost is 3.25566282868385
```

Number 4 appears two time though this maybe have been by accident.

```
inputs:
[|
    [|-1; 1; 1; -1; -1; -1|] // 1
|]

[|
    [|-1; 1; -1; 1; -1; 1|] // 2
|]

[|
    [|-1; 1; 1; 1; 1; 1|] // 3
|]

[|
    [|1; 1; 1; -1; -1; -1|] // 4
|]

[|
    [|-1; 1; 1; -1; -1; 1|] // 5
|]

[|
    [|-1; 1; -1; 1; -1; 1|] // 2
|]

[|
    [|-1; 1; 1; 1; 1; 1|] // 3
|]

[|
    [|-1; 1; 1; -1; -1; 1|] // 5
|]

[|
    [|-1; 1; 1; -1; -1; -1|] // 1
|]

[|
    [|1; 1; 1; -1; -1; -1|] // 4
|]

[|
    [|1; 1; 0; -1; 0; 0|]
|]

target:
[|
    [|1; 1; 1; -1; -1; -1|] // 4
|]
```

That would be it. That 4 repeating twice in the previous thing was just a freak accident. Let me do a thing. Since the vanilla hebb does in fact have biases I should add them.

...No forget it. Later. Right now I need to verify something.

What will happen to the results of the Python net once I set the setting to be the same as in the paper. That is 3 cycles, 10 repetitions and 3 empty inputs?"