``` At iteration 590 the cost is 25 At iteration 600 the cost is 25 At iteration 610 the cost is 25 ```
If I just pass along the strenghened inputs then I get 25 as expected.

3pm.

```
(1999, '====')
[-1.  1.  1. -1. -1. -1. -1.  1.  1. -1.]
[ 20.   0. -20.  -0. -20.  -0.  20.  20.  -0.  20.]
[ 1.    -0.5   -1.    -0.216 -1.    -0.387  1.     1.    -0.143  1.   ]
Time spent on last 10 iters:  0.07811880111694336
Mean loss over last 10 iters: 25.532555198669435
```

This is how a vanilla RNN does on its own. Pretty useless, but it is still on par with random.

```
At iteration 30 the cost is 38.451753616333
At iteration 40 the cost is 36.2502025604248
At iteration 50 the cost is 38.1243562698364
At iteration 60 the cost is 37.6580074310303
At iteration 70 the cost is 35.8993522644043
```

This is how it does in my own library.

...Is it the initialization that is the problem?

I am now suspecting an error on either the forward or the backward pass, but just either of them is so unlikely. Could `activation` be wrong?

Let me set the initialization to the same thing as in the PyTorch.

```
At iteration 80 the cost is 25.2663831710815
At iteration 90 the cost is 25.5244316101074
At iteration 100 the cost is 24.8446262359619
At iteration 110 the cost is 25.6098857879639
At iteration 120 the cost is 25.2827907562256
At iteration 130 the cost is 25.1865591049194
At iteration 140 the cost is 24.8830814361572
```

Ohhh, that son of a bitch. Agh, it got me.

Well, I know what the problem is now. Let me see if this is really the fix to everything.