"3:45pm.
```
The learning rate is 2 ** -12
At iteration 9 the cost is 76.6145587921143
At iteration 19 the cost is 74.1562580108643
At iteration 29 the cost is 64.6177852630615
At iteration 39 the cost is 62.1240394592285
At iteration 49 the cost is 61.4367069244385
At iteration 59 the cost is 62.4488231658936
At iteration 69 the cost is 56.6062015533447
At iteration 79 the cost is 58.2786792755127
At iteration 89 the cost is 53.7622840881348
At iteration 99 the cost is 46.116979598999
At iteration 109 the cost is 51.6013492584229
At iteration 119 the cost is 43.7488735198975
At iteration 129 the cost is 47.0275985717773
At iteration 139 the cost is 43.7086761474609
At iteration 149 the cost is 42.4940631866455
At iteration 159 the cost is 43.7514709472656
At iteration 169 the cost is 44.0407379150391
At iteration 179 the cost is 42.0722875595093
At iteration 189 the cost is 42.741283416748
At iteration 199 the cost is 42.3602062225342
At iteration 209 the cost is 38.9835845947266
At iteration 219 the cost is 40.5237571716309
At iteration 229 the cost is 40.556413269043
At iteration 239 the cost is 42.6226501464844
At iteration 249 the cost is 36.9377471923828
At iteration 259 the cost is 39.4622653961182
At iteration 269 the cost is 39.0161254882812
At iteration 279 the cost is 39.5439588546753
At iteration 289 the cost is 38.8088138580322
At iteration 299 the cost is 40.3712923049927
At iteration 309 the cost is 36.2387546539307
At iteration 319 the cost is 38.0498907089233
At iteration 329 the cost is 39.1001167297363
At iteration 339 the cost is 38.3634624481201
At iteration 349 the cost is 35.8824863433838
At iteration 359 the cost is 39.6173202514648
At iteration 369 the cost is 39.5506057739258
At iteration 379 the cost is 38.2962677001953
At iteration 389 the cost is 40.246125793457
At iteration 399 the cost is 35.0698945999145
At iteration 409 the cost is 32.3450210571289
At iteration 419 the cost is 38.4301158905029
At iteration 429 the cost is 39.1225994110107
At iteration 439 the cost is 37.6665184020996
At iteration 449 the cost is 37.7026443481445
At iteration 459 the cost is 36.2241609573364
At iteration 469 the cost is 37.3382616043091
At iteration 479 the cost is 36.5505294799805
At iteration 489 the cost is 35.7419506072998
At iteration 499 the cost is 34.1014150619507
At iteration 509 the cost is 37.4535085678101
At iteration 519 the cost is 35.5880699157715
At iteration 529 the cost is 34.8883445739746
At iteration 539 the cost is 40.379097366333
At iteration 549 the cost is 38.6012434005737
At iteration 559 the cost is 36.1985210418701
At iteration 569 the cost is 35.561009979248
At iteration 579 the cost is 34.0891284942627
At iteration 589 the cost is 33.9956174850464
At iteration 599 the cost is 38.8431381225586
At iteration 609 the cost is 39.7917011260986
At iteration 619 the cost is 35.7682390213013
At iteration 629 the cost is 36.3878992080688
At iteration 639 the cost is 35.6501062393188
At iteration 649 the cost is 37.1702035903931
At iteration 659 the cost is 35.6945556640625
At iteration 669 the cost is 34.4366394042969
At iteration 679 the cost is 38.4939739227295
At iteration 689 the cost is 38.1441648483276
At iteration 699 the cost is 37.822257232666
At iteration 709 the cost is 34.2027059555054
At iteration 719 the cost is 34.8203477859497
At iteration 729 the cost is 34.5669767379761
At iteration 739 the cost is 36.4086748123169
At iteration 749 the cost is 40.2984085083008
At iteration 759 the cost is 35.7310346603394
At iteration 769 the cost is 36.0016408920288
At iteration 779 the cost is 37.4761524200439
At iteration 789 the cost is 32.357509803772
At iteration 799 the cost is 33.50403175354
At iteration 809 the cost is 34.7146007537842
At iteration 819 the cost is 33.3857765197754
At iteration 829 the cost is 37.8335964202881
At iteration 839 the cost is 34.9598682403564
At iteration 849 the cost is 33.3130405426025
At iteration 859 the cost is 33.158925819397
At iteration 869 the cost is 34.4547340393066
At iteration 879 the cost is 36.7577892303467
At iteration 889 the cost is 37.1376764297485
At iteration 899 the cost is 34.4524425506592
At iteration 909 the cost is 35.0467947006226
At iteration 919 the cost is 34.4010562896729
At iteration 929 the cost is 37.6717569351196
At iteration 939 the cost is 35.1732105255127
At iteration 949 the cost is 33.5719400405884
At iteration 959 the cost is 34.653325843811
At iteration 969 the cost is 37.2138748168945
At iteration 979 the cost is 37.6253520965576
At iteration 989 the cost is 35.4382375717163
At iteration 999 the cost is 37.5381690979004
At iteration 1009 the cost is 32.3703624725342
At iteration 1019 the cost is 33.8052663803101
At iteration 1029 the cost is 37.6619113922119
At iteration 1039 the cost is 36.8373392105103
At iteration 1049 the cost is 40.625150680542
At iteration 1059 the cost is 35.6959976196289
At iteration 1069 the cost is 38.2482261657715
At iteration 1079 the cost is 41.0121315002441
At iteration 1089 the cost is 39.4576686859131
At iteration 1099 the cost is 40.2425163269043
At iteration 1109 the cost is 36.6359231948853
At iteration 1119 the cost is 39.2573657989502
At iteration 1129 the cost is 35.8484107971191
At iteration 1139 the cost is 37.2225269317627
At iteration 1149 the cost is 44.6333778381348
At iteration 1159 the cost is 46.3379228591919
At iteration 1169 the cost is 48.2341026306152
At iteration 1179 the cost is 57.3996486663818
At iteration 1189 the cost is 65.9906604766846
At iteration 1199 the cost is 72.171418762207
At iteration 1209 the cost is 75.1167549133301
At iteration 1219 the cost is 78.5560066223145
At iteration 1229 the cost is 85.5993392944336
At iteration 1239 the cost is 76.2599128723145
At iteration 1249 the cost is 73.7954330444336

Unhandled Exception: ManagedCuda.CudaBlas.CudaBlasException: ExecutionFailed: An access to GPU memory space failed.
```

This is `mi_hebb_prong`.

4:20pm. I am taking a short breather here. I am not sure I feel like doing anything else after that gauntlet. My goal was to get the test to work and now that I've accomplished that my energy is gone. I am thinking what is coming next.

Let me gather my thoughts.

As the curve for the `mi_hebb_prong` demonstrates, it seems I am going to have to abandon the naive notion that plasticity will automatically make things better. I am going to have to go at it one step at a time.

First this specialized MI architecture is going to have to be replaced with a general one which looks more like a vanilla RNN.

What I am going to do from here is make a plastic vanilla RNN. I am going to get rid of that hacky input strengthening. Then I am going to make that work with PRONG. Then I am going to extend that with MI. Not the one I am using now, but the one I used to have in the old library. Then PRONG that one.

Maybe add plasticity to the inputs as well.

The thing is, it is not that my initialization was bad. It is just that the input distribution for this test is much different than for Mnist. Thinking about it a bit more, that probably means that my initialization for the poker game is way too conservative. I should make the weights in the first layer for that bigger. But eering on the side of caution is not bad.

I think that bad initialization is something PRONG could have fixed, but I did not have a vanilla RNN with hebb to try it out on.

4:40pm. Let me take out the input strengthening and I will make a vanilla platic RNN. I'll replace the one I have now as this one is only use for the Binary Pattern test. I do not need it anymore."